Allow standalone sourcemaps
The `sources` list in a source map is supposed to be relative to the map itself. By default, this package creates a list relative to the base path for the files, which is not always relative to the location of the map. Specifically, if each file has its own map, it's likely to be wrong.

This provides a `standalone` option. If `true`, it will use bare, pathless filenames for the paths listed in `sources`.